Rewrite PhysLayoutPane -> PhysLayout
Instead of inheriting javafx.scene.layout.Pane,
the PhysLayout becomes a manager that is instanced
for a given Pane object.

The nodes with connections are tracked, and
only these nodes become a part of the simulation.

The ODE is created whenever the simulation starts.
This will allow adding new nodes and connections
and immediately restarting the simulation.