Make all nodes in the example draggable.
This requires the jfxtras-labs library.

Also, the simulator must now accept position changes
on individual nodes while it is running. The position
is updated and the node's momentum is frozen at zero
until it is "released".

(Currently, this depends on the .pressed() property; maybe add a
way to "freeze" nodes in the PhysLayout - spontaneous mass change?!)