Initial WheelPane implementation.
This has one central node around which the other
children are arranged at equal distances in a certain order.
The central node is fixed, while the other children
can move freely around it.

(No attempt is currently made to dynamically resize children or
to handle children of widely varying sizes.)