Limit the force exerted by force fields at close range.
Within r<.001, the field no longer exerts any force. Within r<.005,
the field exerts the same force toward the center as it does at r=.005.

The movement of a ball at the center of the force field is still erratic,
but it is no longer catapulted out of the field at extreme velocities.