Add new circles and remove old ones interactively.
The last bug shows that layout passes weren't
being tested properly.

This creates new buttons to append or remove a circle,
which triggers a layout pass. As a result, all circles
glide to a new location while the new circle takes its place at the top.

(The circle will first appear in the 0,0 position of the VBox.)